Factor out a file cache from the download cache
The file cache handles persistence of data, etag, and last modified
timestamp. It provides a means to ensure that certain operations are
performed in the file cache only when the associated lock file is
locked exclusively and it delivers the cached data after downgrading
the exclusive lock to a shared lock.

This lock downgrade is actually pointless as two processes cannot
share the lock if the only way to the shared lock is via a downgrade.

[#59478422]